requirements.txt
wheel
Django==4.0.2
requests
djangorestframework==3.13.1
mysqlclient